requirements.txt
langchain>=0.1.0
openai>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
pandas>=2.0.0
scikit-learn
aiohttp>=3.8.0
python-dotenv>=1.0.0
numpy>=1.20.0
pandas
python-dotenv
langchain_groq
sentence_transformers